Move visible lines to custom hook
